# packages/contracts/openapi/openapi.generated.json
{
  "components": {
    "schemas": {
...
      "AdminProductDetailResponse": {
        "properties": {
          "archivedAt": {
            "description": "Set only for an archived product.",
            "format": "date-time",
            "type": "string"
          },
          "basePriceAmount": {
            "description": "Whole đồng as a decimal string; never a JSON number. `0` on a DRAFT means the price has not been set yet.",
            "example": "0",
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/AdminProductCategoryResponse"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "currencyCode": {
            "enum": [
              "VND"
            ],
            "example": "VND",
            "type": "string"
          },
          "description": {
            "description": "Absent when the draft has no description. On PATCH, sending null or a blank string clears it; omitting the field leaves it unchanged.",
            "type": "string"
          },
          "media": {
            "description": "Ordered media selection.",
            "items": {
              "$ref": "#/components/schemas/AdminProductMediaResponse"
            },
            "type": "array"
          },
          "name": {
            "example": "Thú bông gấu nâu",
            "type": "string"
          },
          "primaryMedia": {
            "allOf": [
              {
                "$ref": "#/components/schemas/AdminProductMediaResponse"
              }
            ],
            "description": "The first selected image, when one exists. Identity only — there is no URL."
          },
          "productId": {
            "example": "019a2b3c-4d5e-7f60-8a1b-2c3d4e5f6071",
            "format": "uuid",
            "type": "string"
          },
          "slug": {
            "description": "Server-owned public address. Immutable in APP2-B02.",
            "example": "thu-bong-gau-nau",
            "type": "string"
          },
          "status": {
            "enum": [
              "DRAFT",
              "PUBLISHED",
              "ARCHIVED"
            ],
            "example": "DRAFT",
            "type": "string"
          },
          "updatedAt": {
            "description": "Optimistic-concurrency token; send it back as `expectedUpdatedAt`.",
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "productId",
          "name",
          "slug",
          "status",
          "category",
          "basePriceAmount",
          "currencyCode",
          "createdAt",
          "updatedAt",
          "media"
        ],
        "type": "object"
      },
...
      "ApiErrorResponse": {
        "description": "Canonical error envelope (D-034). Server faults always carry the generic message; internal detail is never published.",
        "properties": {
          "code": {
            "description": "Stable machine-readable code; branch on this.",
            "enum": [
              "BAD_REQUEST",
              "UNAUTHORIZED",
              "FORBIDDEN",
              "NOT_FOUND",
              "METHOD_NOT_ALLOWED",
              "CONFLICT",
              "UNPROCESSABLE_ENTITY",
              "UNSUPPORTED_MEDIA_TYPE",
              "TOO_MANY_REQUESTS",
              "INTERNAL_SERVER_ERROR"
            ],
            "type": "string"
          },
          "errors": {
            "description": "Structured validation or domain details, when safe to publish.",
            "items": {
              "$ref": "#/components/schemas/ApiFieldError"
            },
            "type": "array"
          },
          "message": {
            "description": "Human-readable text; never branch on this.",
            "type": "string"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiResponseMeta"
          },
          "success": {
            "enum": [
              false
            ],
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "code",
          "message",
          "meta"
        ],
        "type": "object"
      },
...
      "ApiSuccessResponse": {
        "description": "Canonical success envelope (D-034). Endpoint responses compose this with their own payload schema for `data`.",
        "properties": {
          "code": {
            "description": "Stable machine-readable code; branch on this.",
            "type": "string"
          },
          "data": {
            "description": "Endpoint payload; null where the contract has no payload.",
            "nullable": true
          },
          "message": {
            "description": "Human-readable text; never branch on this.",
            "type": "string"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiResponseMeta"
          },
          "success": {
            "enum": [
              true
            ],
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "code",
          "message",
          "data",
          "meta"
        ],
        "type": "object"
      },
...
      "UpdateProductBody": {
        "properties": {},
        "type": "object"
      }
    },
    "securitySchemes": {
      "adminSession": {
        "in": "cookie",
        "name": "adm_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Internal HTTP contract for the Embroidery Commerce platform API. Generated from NestJS Swagger metadata; the committed artifact is the machine-readable source for the generated TypeScript/Axios client.",
    "title": "Embroidery Commerce API",
    "version": "0.1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/products/{productId}": {
      "get": {
        "description": "Returns the editable draft fields, the ordered media selection and the `updatedAt` concurrency token.",
        "operationId": "adminProduct_detail",
        "parameters": [
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "schema": {
              "format": "uuid"
            }
          },
          {
            "description": "Optional client-supplied correlation ID. A value matching the pattern is preserved end to end; anything missing or outside it is replaced by a safe generated ID and is never echoed back.",
            "in": "header",
            "name": "X-Request-ID",
            "required": false,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9._-]{1,64}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/AdminProductDetailResponse"
                        }
                      },
                      "required": [
                        "data"
                      ],
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "The product.",
            "headers": {
              "X-Request-ID": {
                "description": "The effective correlation ID for this request, set by the gateway. Quote it when reporting a problem so the request can be found in the logs.",
                "schema": {
                  "maxLength": 64,
                  "minLength": 1,
                  "pattern": "^[A-Za-z0-9._-]{1,64}$",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "No live Admin session.",
            "headers": {
              "X-Request-ID": {
                "description": "The effective correlation ID for this request, set by the gateway. Quote it when reporting a problem so the request can be found in the logs.",
                "schema": {
                  "maxLength": 64,
                  "minLength": 1,
                  "pattern": "^[A-Za-z0-9._-]{1,64}$",
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "No such product.",
            "headers": {
              "X-Request-ID": {
                "description": "The effective correlation ID for this request, set by the gateway. Quote it when reporting a problem so the request can be found in the logs.",
                "schema": {
                  "maxLength": 64,
                  "minLength": 1,
                  "pattern": "^[A-Za-z0-9._-]{1,64}$",
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "An unexpected error occurred. Please try again later.",
            "headers": {
              "X-Request-ID": {
                "description": "The effective correlation ID for this request, set by the gateway. Quote it when reporting a problem so the request can be found in the logs.",
                "schema": {
                  "maxLength": 64,
                  "minLength": 1,
                  "pattern": "^[A-Za-z0-9._-]{1,64}$",
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "Get one Admin product",
        "tags": [
          "adminProduct"
        ]
      },
      "patch": {
        "description": "Patches editable DRAFT fields and, when `mediaAssetIds` is present, replaces the whole ordered media selection. Requires `expectedUpdatedAt`; a stale value is rejected as a conflict rather than overwriting a concurrent change. Renaming never changes the slug.",
        "operationId": "adminProduct_update",
        "parameters": [
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "schema": {
              "format": "uuid"
            }
          },
          {
            "description": "Optional client-supplied correlation ID. A value matching the pattern is preserved end to end; anything missing or outside it is replaced by a safe generated ID and is never echoed back.",
            "in": "header",
            "name": "X-Request-ID",
            "required": false,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9._-]{1,64}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/AdminProductDetailResponse"
                        }
                      },
                      "required": [
                        "data"
                      ],
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "The updated draft.",
            "headers": {
              "X-Request-ID": {
                "description": "The effective correlation ID for this request, set by the gateway. Quote it when reporting a problem so the request can be found in the logs.",
                "schema": {
                  "maxLength": 64,
                  "minLength": 1,
                  "pattern": "^[A-Za-z0-9._-]{1,64}$",
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Invalid body, category or media.",
            "headers": {
              "X-Request-ID": {
                "description": "The effective correlation ID for this request, set by the gateway. Quote it when reporting a problem so the request can be found in the logs.",
                "schema": {
                  "maxLength": 64,
                  "minLength": 1,
                  "pattern": "^[A-Za-z0-9._-]{1,64}$",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "No live Admin session.",
            "headers": {
              "X-Request-ID": {
                "description": "The effective correlation ID for this request, set by the gateway. Quote it when reporting a problem so the request can be found in the logs.",
                "schema": {
                  "maxLength": 64,
                  "minLength": 1,
                  "pattern": "^[A-Za-z0-9._-]{1,64}$",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "The request states an origin outside the Admin allowlist.",
            "headers": {
              "X-Request-ID": {
                "description": "The effective correlation ID for this request, set by the gateway. Quote it when reporting a problem so the request can be found in the logs.",
                "schema": {
                  "maxLength": 64,
                  "minLength": 1,
                  "pattern": "^[A-Za-z0-9._-]{1,64}$",
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "No such product.",
            "headers": {
              "X-Request-ID": {
                "description": "The effective correlation ID for this request, set by the gateway. Quote it when reporting a problem so the request can be found in the logs.",
                "schema": {
                  "maxLength": 64,
                  "minLength": 1,
                  "pattern": "^[A-Za-z0-9._-]{1,64}$",
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Stale `expectedUpdatedAt`, a non-editable state, or an unusable image.",
            "headers": {
              "X-Request-ID": {
                "description": "The effective correlation ID for this request, set by the gateway. Quote it when reporting a problem so the request can be found in the logs.",
                "schema": {
                  "maxLength": 64,
                  "minLength": 1,
                  "pattern": "^[A-Za-z0-9._-]{1,64}$",
                  "type": "string"
                }
              }
            }
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "The body is not application/json.",
            "headers": {
              "X-Request-ID": {
                "description": "The effective correlation ID for this request, set by the gateway. Quote it when reporting a problem so the request can be found in the logs.",
                "schema": {
                  "maxLength": 64,
                  "minLength": 1,
                  "pattern": "^[A-Za-z0-9._-]{1,64}$",
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "An unexpected error occurred. Please try again later.",
            "headers": {
              "X-Request-ID": {
                "description": "The effective correlation ID for this request, set by the gateway. Quote it when reporting a problem so the request can be found in the logs.",
                "schema": {
                  "maxLength": 64,
                  "minLength": 1,
                  "pattern": "^[A-Za-z0-9._-]{1,64}$",
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "Update a product draft",
        "tags": [
          "adminProduct"
        ]
      }
    },
...
    "/api/public/products": {
      "get": {
        "description": "Returns published products in editorial order, page by page. Anonymous: no session or cookie is involved, and the caller cannot select lifecycle visibility — there is no parameter for it, and drafts and archived products are excluded by the query itself. Pagination is keyset: `nextCursor` is opaque, bound to the filter it was issued under, and null on the last page. Responses are never stored. Publication is re-read on every request, and there is no cache-invalidation consumer in this system, so a stored copy could keep an unpublished product visible.",
        "operationId": "publicProduct_list",
        "parameters": [
          {
            "in": "query",
            "name": "categorySlug",
            "required": false,
            "schema": {
              "enum": [
                "thu-bong",
                "khan",
                "quan-ao",
                "khac"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "description": "Opaque cursor from a prior page.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {}
          },
          {
            "description": "Optional client-supplied correlation ID. A value matching the pattern is preserved end to end; anything missing or outside it is replaced by a safe generated ID and is never echoed back.",
            "in": "header",
            "name": "X-Request-ID",
            "required": false,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9._-]{1,64}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PublicProductListResponse"
                        }
                      },
                      "required": [
                        "data"
                      ],
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "A page of published products.",
            "headers": {
              "X-Request-ID": {
                "description": "The effective correlation ID for this request, set by the gateway. Quote it when reporting a problem so the request can be found in the logs.",
                "schema": {
                  "maxLength": 64,
                  "minLength": 1,
                  "pattern": "^[A-Za-z0-9._-]{1,64}$",
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "An unknown query parameter, an out-of-range page size, or a cursor that is malformed or was issued for a different filter.",
            "headers": {
              "X-Request-ID": {
                "description": "The effective correlation ID for this request, set by the gateway. Quote it when reporting a problem so the request can be found in the logs.",
                "schema": {
                  "maxLength": 64,
                  "minLength": 1,
                  "pattern": "^[A-Za-z0-9._-]{1,64}$",
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "An unexpected error occurred. Please try again later.",
            "headers": {
              "X-Request-ID": {
                "description": "The effective correlation ID for this request, set by the gateway. Quote it when reporting a problem so the request can be found in the logs.",
                "schema": {
                  "maxLength": 64,
                  "minLength": 1,
                  "pattern": "^[A-Za-z0-9._-]{1,64}$",
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "List published products",
        "tags": [
          "publicProduct"
        ]
      }
    },
    "/api/public/products/{slug}": {
      "get": {
        "description": "Resolves a published product by its immutable server-owned slug. An unknown slug, a draft, an archived product and a product without a public category all return the same 404: a public caller must not be able to tell unreleased work from work that never existed. Responses are never stored. Publication is re-read on every request, and there is no cache-invalidation consumer in this system, so a stored copy could keep an unpublished product visible.",
        "operationId": "publicProduct_detail",
        "parameters": [
          {
            "description": "The immutable server-owned product slug.",
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {}
          },
          {
            "description": "Optional client-supplied correlation ID. A value matching the pattern is preserved end to end; anything missing or outside it is replaced by a safe generated ID and is never echoed back.",
            "in": "header",
            "name": "X-Request-ID",
            "required": false,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9._-]{1,64}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessResponse"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PublicProductDetailResponse"
                        }
                      },
                      "required": [
                        "data"
                      ],
                      "type": "object"
                    }
                  ]
                }
              }
            },
            "description": "The published product.",
            "headers": {
              "X-Request-ID": {
                "description": "The effective correlation ID for this request, set by the gateway. Quote it when reporting a problem so the request can be found in the logs.",
                "schema": {
                  "maxLength": 64,
                  "minLength": 1,
                  "pattern": "^[A-Za-z0-9._-]{1,64}$",
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "The slug is not well formed.",
            "headers": {
              "X-Request-ID": {
                "description": "The effective correlation ID for this request, set by the gateway. Quote it when reporting a problem so the request can be found in the logs.",
                "schema": {
                  "maxLength": 64,
                  "minLength": 1,
                  "pattern": "^[A-Za-z0-9._-]{1,64}$",
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "No published product is available at that slug.",
            "headers": {
              "X-Request-ID": {
                "description": "The effective correlation ID for this request, set by the gateway. Quote it when reporting a problem so the request can be found in the logs.",
                "schema": {
                  "maxLength": 64,
                  "minLength": 1,
                  "pattern": "^[A-Za-z0-9._-]{1,64}$",
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "An unexpected error occurred. Please try again later.",
            "headers": {
              "X-Request-ID": {
                "description": "The effective correlation ID for this request, set by the gateway. Quote it when reporting a problem so the request can be found in the logs.",
                "schema": {
                  "maxLength": 64,
                  "minLength": 1,
                  "pattern": "^[A-Za-z0-9._-]{1,64}$",
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Get one published product by slug",
        "tags": [
          "publicProduct"
        ]
      }
    },
...
  },
  "servers": [],
...
}